chore: restructure and rename widgts
