+++ lib/widgets/content/tags/section.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import '../../../i18n/_generated_/translations.g.dart';
 import '../../../models/book/content/section_tag.dart';
 import '../../../types/types.dart';
-import '../../book/get_tag_list.dart';
+import '../../helpers/get_tag_list.dart';
 import '../../typography/headline.dart';
 
 class Section extends StatelessWidget {
