+++ lib/widgets/matter/footnotes.dart
@@ -1,18 +1,17 @@
 import 'package:flutter/material.dart';
 
-import '../../models/book/content/section_tag.dart';
 import '../../models/book/section/footnote.dart';
 import '../../types/types.dart';
 import '../content/list/footnotes_list_item.dart';
 import '../layout/content_container.dart';
 import '../typography/headline.dart';
 
-class FootnotesList extends StatelessWidget {
+class Footnotes extends StatelessWidget {
   final List<Footnote> footnotes;
   final OnNavigate onNavigate;
-  final SectionTag footnoteSection;
+  final BookText sectionTitle;
 
-  const FootnotesList(this.footnoteSection, this.footnotes, this.onNavigate, {Key? key}) : super(key: key);
+  const Footnotes(this.sectionTitle, this.onNavigate, this.footnotes, {Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -28,7 +27,7 @@ class FootnotesList extends StatelessWidget {
       child: Column(
         mainAxisAlignment: MainAxisAlignment.start,
         children: [
-          Headline(footnoteSection.meta.title),
+          Headline(sectionTitle),
           ...footnoteItems,
         ],
       ),
